+++ src/ClassicUO.Client/Game/World.cs
@@ -351,6 +351,7 @@ namespace ClassicUO.Game
             {
                 string lastChar = LastCharacterManager.GetLastCharacter(LoginScene.Account, ServerName);
                 ProfileManager.Load(ServerName, LoginScene.Account, lastChar, serial);
+                SoundFilterManager.MigrateLegacySqlSettings();
             }
 
             if (Player != null)