+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to TazUO will be recorded here.
 ### Features
 * Separated Scavenger agent from Autoloot, they now each have their own loot lists and enabled/disabled toggles
 
+### Misc
+* Migrated more settings to global scoped json settings
+
 ### Fixes
 * Myra sliders and checkboxes no longer listen to right clicks (Causing accidental changed when closing via right click over a checkbox/slider)
 * Improved treasure map location calculations - [P.R 1012](https://github.com/PlayTazUO/TazUO/pull/1012) [Erumite](https://github.com/Erumite)